authentication

+++ base/views.py
@@ -1,11 +1,24 @@
 from django.shortcuts import render,redirect
+from django.db.models import Q
+from django.contrib.auth.models import User
+from django.contrib import messages
 from .models import *
+from django.contrib.auth import authenticate,login,logout
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.forms import UserCreationForm
 from .forms import RoomForm
 
 # Create your views here.
 def home(request):
-    rooms = Room.objects.all()
-    context = {"rooms":rooms}
+    q = request.GET.get('q') if request.GET.get('q')!=None else ''
+    topics = Topic.objects.all()
+    rooms = Room.objects.filter(Q(topic__name__icontains=q) | Q(description__icontains=q) | Q(name__icontains=q))
+    room_count = rooms.count()
+    context = {
+        "rooms":rooms,
+        "topics":topics,
+        "room_count":room_count,
+        }
     return render(request,'base/home.html',context)
 
 def room(request,pk):
@@ -13,6 +26,7 @@ def room(request,pk):
     context = {"room":room}
     return render(request,'base/room.html',context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -26,8 +40,12 @@ def createRoom(request):
     }
     return render(request,'base/room_form.html',context)
 
+@login_required(login_url='login')
 def updateRoom(request,pk):
     room = Room.objects.get(id=pk)
+    if request.user!=room.host:
+        messages.error(request,'Access Denied')
+        return redirect('home')
     form = RoomForm(instance=room)
     if request.method == 'POST':
         form = RoomForm(request.POST,instance=room)
@@ -39,11 +57,60 @@ def updateRoom(request,pk):
     }
     return render(request,'base/room_form.html',context)
 
+@login_required(login_url='login')
 def deleteRoom(request,pk):
     room = Room.objects.get(id=pk)
+    if request.user!=room.host:
+        messages.error(request,'Access Denied')
+        return redirect('home')
+
     if request.method=='POST':
         room.delete()
         return redirect('home')
     context = {'obj':room}
     return render(request,'base/delete.html',context)
     
+def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
+    if request.method=="POST":
+        username = request.POST.get('username').lower()
+        password = request.POST.get('password')
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request,"User Does Not Exists")
+        user = authenticate(request,username=username,password=password)
+        if user is not None:
+            login(request,user)
+            return redirect('home')
+        else:
+            messages.error(request,"Invalid Credentials")
+    context = {"page":page}
+    return render(request,'base/login_register.html',context)
+
+def registerPage(request):
+    form = UserCreationForm()
+    if request.method=="POST":
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request,user)
+            messages.info(request,'User has been registered!')
+            return redirect('home')
+        else:
+            messages.error(request,'Error Occurred During Registration')
+            return redirect('')
+    context = {
+        "form":form}
+    return render(request,'base/login_register.html',context)
+
+
+@login_required(login_url='login')
+def logoutPage(request):
+    logout(request)
+    messages.info(request,"Successfully Logged Out")
+    return redirect('home')
